requirements.txt
asgiref==3.3.1
certifi==2020.11.8
Django==3.0.8
django-filter==2.4.0
django-tinymce4-lite==1.8.0
gunicorn==20.1.0
jsmin==2.2.2
Pillow==8.3.1
psycopg2-binary==2.9.1
pytz==2021.1
sqlparse==0.4.1
